requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
discord.py==1.3.1
idna==2.8
multidict==4.5.2
requests==2.22.0
urllib3==1.25.6
websockets==6.0
yarl==1.3.0
